Move selectors out of deleted file
