+++ src/selectors/Bluetooth/bluetooth.js
@@ -1,3 +1,8 @@
+/**
+ * mSupply Mobile
+ * Sustainable Solutions (NZ) Ltd. 2021
+ */
+
 export const selectSendingBlinkTo = ({ bluetooth }) => {
   const { blink } = bluetooth || {};
   const { sendingBlinkTo = '' } = blink || {};
@@ -16,8 +21,35 @@ export const selectScannedSensors = ({ bluetooth }) => {
   return scannedSensorAddresses;
 };
 
-export const selectIsDownloadingLogs = ({ bluetooth }) => {
+export const selectIsSyncingTemps = ({ bluetooth }) => {
   const { download } = bluetooth || {};
-  const { isDownloadingLogs = false } = download || {};
-  return isDownloadingLogs;
+  const { isSyncingTemps = false } = download || {};
+  return isSyncingTemps;
+};
+
+export const selectDownloadingLogsFrom = ({ bluetooth }) => {
+  const { download } = bluetooth || {};
+  const { downloadingLogsFrom } = download;
+  return downloadingLogsFrom;
+};
+
+export const selectLastDownloadTime = ({ bluetooth }, macAddress) => {
+  const { download } = bluetooth || {};
+  const { lastDownloadTime } = download;
+  return lastDownloadTime[macAddress] ? new Date(lastDownloadTime[macAddress]) : null;
+};
+
+export const selectLastDownloadFailed = ({ bluetooth }, macAddress) => {
+  const { download } = bluetooth || {};
+  const { lastDownloadStatus } = download;
+  const status = lastDownloadStatus[macAddress];
+  const lastDownloadFailed = status !== 'OK';
+  return lastDownloadFailed;
+};
+
+export const selectIsDownloading = (state, macAddress) => {
+  const downloadingLogsFrom = selectDownloadingLogsFrom(state);
+  const isDownloading = downloadingLogsFrom === macAddress;
+
+  return isDownloading;
 };
