+++ src/widgets/LastSensorDownload.js
@@ -14,7 +14,7 @@ import {
   selectIsDownloading,
   selectLastDownloadFailed,
   selectLastDownloadTime,
-} from '../selectors/vaccine';
+} from '../selectors/Bluetooth/bluetooth';
 import { generalStrings } from '../localization';
 
 const formatLastSyncDate = date => (date ? moment(date).fromNow() : generalStrings.not_available);
